error handling + bug fixing + minor UI changes
