+++ frontend/src/components/message/botMessage.tsx
@@ -1,74 +1,50 @@
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faRotateRight, faChevronDown, faChevronUp } from '@fortawesome/free-solid-svg-icons';
+import type { splittedMessage } from "../../api/types";
+import { useState } from "react";
 import ReactMarkdown from 'react-markdown';
 import { CodeBlock } from '../codeBlock';
 import remarkGfm from 'remark-gfm';
 import remarkMath from 'remark-math';
 import rehypeKatex from 'rehype-katex';
-import { useEffect, useState } from 'react';
 import type { ComponentPropsWithoutRef } from 'react';
-
-
-const ThinkingPlaceholder = () => (
-    <span className="font-semibold relative inline-block shimmer-text text-transparent bg-clip-text italic">Please wait, I'm thinking...</span>
-);
-
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faRotateRight, faChevronDown, faChevronUp } from '@fortawesome/free-solid-svg-icons';
+import { useConversationLogic } from '../../hooks/useConversationLogic';
 
 type CodeComponentProps = ComponentPropsWithoutRef<'code'> & {
     inline?: boolean;
 };
 
-export function BotMessage({
-    message,
-    onReload,
-    token,
-    think }: {
-        message: string | null;
-        onReload: () => void;
-        token: number,
-        think?: string
-    }) {
+
+export function BotMessage({ splittedMessage, isMulti, isLast, setError }: { splittedMessage: splittedMessage, isMulti?: boolean, isLast?: boolean, setError?: (error: string | null) => void }) {
 
     const [showThinking, setShowThinking] = useState(false);
-    const [isLoading, setIsLoading] = useState(message === null || message.trim() === "");
+    const { regenerateMessage, chooseReply } = useConversationLogic();
 
+    const msgIsNotEmpty = splittedMessage.mainContent.trim() !== '' || splittedMessage.thinkContent.trim() !== '';
 
-    useEffect(() => {
-        if (message && message.trim() !== "") {
-            setIsLoading(false);
-        } else {
-            setIsLoading(true);
+    const handleChooseReply = async () => {
+        try {
+            await chooseReply(splittedMessage.convId, splittedMessage.msgId);
+        } catch (error) {
+            console.log("should show modal with error : ", error);
+            setError?.(error instanceof Error ? error.message : 'An unexpected error occurred');
         }
-    }, [message]);
+    }
 
 
     return (
-
-        <div className="self-start ml-0 md:ml-15 mt-5 w-fit max-w-[90%]">
-            {think && (
-                <div className="mb-2 text-xs text-gray-100">
-                    <button
-                        className="flex items-center gap-1 hover:cursor-pointer focus:outline-none"
-                        onClick={() => setShowThinking(!showThinking)}
-                    >
-                        <span
-                            className={`font-semibold relative inline-block ${isLoading ? 'shimmer-text text-transparent bg-clip-text' : ''
-                                }`}
-                        >
-                            {isLoading ? 'Thinking...' : 'Thinking'}
-                        </span>
-                        <FontAwesomeIcon icon={showThinking ? faChevronUp : faChevronDown} size="xs" />
-                    </button>
-
-                    <div className={`transition-all duration-300 ease-in-out ${showThinking ? 'opacity-100 mt-1' : 'max-h-0 opacity-0'} overflow-hidden`}>
-                        <div className="p-2 rounded-md bg-blue-900/20 text-blue-300 text-sm whitespace-pre-wrap font-mono">
-                            {think}
-                        </div>
+        <div key={splittedMessage.msgId} className={`flex flex-col py-4 gap-2 ${isMulti ? 'bg-white/1 rounded-lg px-8' : 'px-4 md:px-40'}`}>
+            {isMulti && <h2 className="pb-2 border-b border-[var(--color-separator)] text-center">{splittedMessage.author}</h2>}
+            <div className={'pb-2 whitespace-pre-wrap flex-1}'}>
+                {splittedMessage.thinkContent && (<button className="bg-white/5 px-4 py-2 my-5 rounded-lg" onClick={() => setShowThinking(!showThinking)}>{
+                    showThinking ? <FontAwesomeIcon icon={faChevronUp} size="sm" /> : <FontAwesomeIcon icon={faChevronDown} size="sm" />
+                } Thinking</button>)}
+                {showThinking && splittedMessage.thinkContent && (
+                    <div className="text-blue-300 mb-2 thinkingText">
+                        {splittedMessage.thinkContent}
                     </div>
-                </div>
-            )}
-            <div className="prose prose-invert max-w-none text-gray-300 text-sm p-4">
-                {!isLoading ? (
+                )}
+                <div className={`hide-scrollbar overflow-y-auto ${isMulti ? 'max-h-[50vh]' : ''}`}>
                     <ReactMarkdown
                         remarkPlugins={[remarkGfm, remarkMath]}
                         rehypePlugins={[rehypeKatex]}
@@ -82,37 +58,54 @@ export function BotMessage({
                                     return <CodeBlock language={match[1]} value={codeContent} />;
                                 }
 
-                                return <code className="bg-gray-800 px-1 rounded">{children}</code>;
+                                return <code>{children}</code>;
                             },
                         }}
 
                     >
-                        {message}
+                        {splittedMessage.mainContent}
                     </ReactMarkdown>
-                ) : (
-                    // If the message already has think content we don't need to say it again
-                    think ? null : <ThinkingPlaceholder />
-                )}
+                </div>
             </div>
-
-            <div className="border-t-2 border-gray-800 mb-2 w-full" />
-
-            {token > 0 ? (
-                <div className="flex justify-between items-center text-[10px] text-gray-500 italic w-full px-1">
-
-                    <span>{token} tokens</span>
-                    <button onClick={onReload} className="text-gray-500 hover:text-gray-300 transition-colors cursor-pointer">
-                        <FontAwesomeIcon icon={faRotateRight} />
-                    </button>
+            <div className="smallText pt-2 border-t border-[var(--color-separator)] mt-auto">
+                <div className="flex">
+                    <div
+                        className="flex-1 flex gap-5"
+                    >
+                        <span>
+                            {splittedMessage.token || 'N/A'} Tokens
+                        </span>
+                        <span>
+                            {new Date(splittedMessage.timestamp).toLocaleDateString('fr-FR', {
+                                year: 'numeric',
+                                month: '2-digit',
+                                day: '2-digit',
+                                hour: '2-digit',
+                                minute: '2-digit',
+                                second: '2-digit',
+                            })
+                            }
+                        </span>
+                    </div>
+                    <div
+                        className="text-right"
+                    >
+                        <button
+                            className="text-gray-400 hover:text-white transition-colors mr-1 cursor-pointer"
+                            onClick={() => {
+                                navigator.clipboard.writeText(splittedMessage.mainContent);
+                            }}
+                        >
+                            <FontAwesomeIcon icon={faRotateRight} onClick={() => regenerateMessage(splittedMessage.convId, splittedMessage.msgId, splittedMessage.author)} />
+                        </button>
+                    </div>
                 </div>
-            ) : (
-                <div className="flex justify-end items-center text-[10px] text-gray-500 italic w-full px-1">
-                    <button onClick={onReload} className="text-gray-500 hover:text-gray-300 transition-colors cursor-pointer">
-                        <FontAwesomeIcon icon={faRotateRight} />
-                    </button>
-                </div>
-            )
-            }
+            </div>
+            {isLast && isMulti && msgIsNotEmpty && (
+                <button className="bg-blue-500 text-white px-4 py-2 rounded-lg shadow hover:bg-blue-600 cursor-pointer transition-colors duration-200 w-fit mx-auto text-md" onClick={handleChooseReply}>
+                    Continue from this answer
+                </button>
+            )}
         </div>
-    );
+    )
 }
