+++ frontend/src/index.css
@@ -4,8 +4,8 @@
 
 
 :root {
-  --color-bg: hsl(0, 1%, 11%);
---color-onTop: hsla(0, 0%, 100%, 0.05);
+  --color-bg: hsl(228, 9%, 10%);
+--color-onTop: hsla(228, 20%, 65%, 0.097);
 --color-separator: hsla(0, 0%, 100%, 0.1);
 --color-CTA: hsla(90, 90%, 50%, 1);
 --color-regularTextColor: hsla(0, 0%, 100%, 1);
@@ -15,6 +15,7 @@
 
 @theme {
 --font-playfair: 'Playfair Display', serif;
+--font-montserrat: 'Montserrat', sans-serif;
 --font-raleway: 'Raleway', sans-serif;
 }
